Moved obsolete methods to ResponseBuilderLegacy
